edje player - fix ridiculous use of srncat
another case of blind strncat usage rthat seemingly did a:

strcat(buf, str); -> strncat(buf, str, strlen(tr));

which is the most pointless thing to do ever. it's no better. it's
worse as it makes it harder to identify issues. thanks these days
compilers warn about this as a stupid thing to do... so i looked at it
and fixed it properly.

again - don't do this stuff above - it's pointless. patches that do
this should be rejected. fix it properly or leave it as-is.